Fix cli info format
Currently the convert script always displays the total number of formats
it is able to export to, no matter how many of which have actually been
selected:
```
Converted data for 250 countries into 5 formats.
```
This is now fixed. Calling the convert script with, say, `-f csv` it now
displays `[...] into 1 format.`